add timerTextModifier to list of modifiers
