+++ Bombus/View/Honeycomb/HoneycombPolygon+Extension.swift
