+++ Bombus/View/SessionView/SessionView.swift
@@ -11,24 +11,11 @@ struct SessionView: View {
             VStack {
                 
                 ZStack {
-                    HoneycombPolygon(radius: 20).foregroundColor(.bombusbAccentColor)
                     
-                    HoneycombPolygon(radius: 20)
-                        .mask(
-                            Color.bombusPrimaryTextColor
-                                .mask(
-                                    HoneycombPolygon(radius: 20)
-                                        .strokeBorder(Color.black.opacity(0.7), lineWidth: 3)
-                                        .blur(radius: 15)
-                                    
-                                    
-                                )
-                            )
+                    Honeycomb()
+                    
                     Text("\(viewModel.formatter.string(from: viewModel.timerDuration) ?? "Nothing")")
-                        .font(.system(size: 60, weight: .bold, design: .rounded))
-                        .fontWeight(.bold)
-                        .padding(50)
-                        .foregroundColor(.bombusPrimaryTextColor)
+                        .timerTextStyle()
                     
                 }
                 .frame(height: 320)
@@ -59,42 +46,5 @@ struct SessionView_Previews: PreviewProvider {
         dummySession.start = Date.init(timeIntervalSinceNow: -10)
         
         return SessionView(storageProvider: viewContext)
-    }
-}
-
-extension View {
-    func innerShadow(color: Color, radius: CGFloat = 0.1) -> some View {
-        modifier(InnerShadow(color: color, radius: min(max(0, radius), 1)))
-    }
-}
-
-private struct InnerShadow: ViewModifier {
-    var color: Color = .gray
-    var radius: CGFloat = 0.1
-    
-    private var colors: [Color] {
-        [color.opacity(0.75), color.opacity(0.0), .clear]
-    }
-    
-    func body(content: Content) -> some View {
-        GeometryReader { geo in
-            content
-                .overlay(LinearGradient(gradient: Gradient(colors: self.colors), startPoint: .top, endPoint: .bottom)
-                            .frame(height: self.radius * self.minSide(geo)),
-                         alignment: .top)
-                .overlay(LinearGradient(gradient: Gradient(colors: self.colors), startPoint: .bottom, endPoint: .top)
-                            .frame(height: self.radius * self.minSide(geo)),
-                         alignment: .bottom)
-                .overlay(LinearGradient(gradient: Gradient(colors: self.colors), startPoint: .leading, endPoint: .trailing)
-                            .frame(width: self.radius * self.minSide(geo)),
-                         alignment: .leading)
-                .overlay(LinearGradient(gradient: Gradient(colors: self.colors), startPoint: .trailing, endPoint: .leading)
-                            .frame(width: self.radius * self.minSide(geo)),
-                         alignment: .trailing)
-        }
-    }
-    
-    func minSide(_ geo: GeometryProxy) -> CGFloat {
-        CGFloat(3) * min(geo.size.width, geo.size.height) / 2
     }
 }
